reduce test cases
